requirements.txt
flask==2.0.2
lint==1.2.1
pylint==2.11.1
pyyaml==5.4.1
flask-restx==0.5.1
pandas==1.3.4
flask-expects-json==1.6.0
scikit-learn==0.24.2
sklearn2pmml==0.77.1